_SectionsReorg_ service added
Public service _DctAddAscByKey_ added which may be copied instead of importing the whole Common Component. It provides what is probably most commonly used setup: Add in ascending order by key (when key is an object its Name property is used as the sort criteria, if it doesn't have a Name property the order is unpredictable.